requirements.txt
# Aetheer v3.0 — dependencias Python
# v1.2 legacy (MCP servers, scraping fallback, Pine parser) se mantiene funcionando.
# Nuevas deps para la capa OpenRouter + orquestador (Fase 1 en adelante).

# --- v1.2 legacy (no tocar salvo bump de seguridad) ---
httpx>=0.27.0
aiohttp>=3.9.0
feedparser>=6.0.0
python-dateutil>=2.9.0
pyyaml>=6.0.0
mcp>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0

# --- v3.0 orquestador (Fase 1) ---
# OpenRouterClient depende solo de httpx (ya presente) para POST async.
# pydantic se añadirá en Fase 2 cuando reemplacemos los JSON Schemas manuales.
pydantic>=2.8.0

# --- v3.0 scheduler (Fase 5) ---
apscheduler>=3.11.0

# --- v3.0 interfaces (Fase 6) ---
fastapi>=0.115.0
uvicorn>=0.30.0
python-telegram-bot>=21.5
redis>=5.0.8

# --- v3.0 load tests (Fase 7) ---
locust>=2.31.0

# --- Testing (dev) ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
freezegun>=1.5.0
